pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-warehouse-env"
version = "0.1.0"
description = "Warehouse Order Fulfillment Environment for OpenEnv - agent navigates warehouse to pick items and deliver orders"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv dependencies
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    # Environment-specific
    "fastmcp>=0.1.0",
    # Inference script
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "warehouse_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["warehouse_env", "warehouse_env.server"]
package-dir = { "warehouse_env" = ".", "warehouse_env.server" = "server" }
